begin great move to notcurses
