+++ debian/control
@@ -13,10 +13,9 @@ Build-Depends:
  libbluetooth-dev,
  libcap-dev,
  libiw-dev,
- libncurses-dev (>= 6.1),
  libnl-3-dev,
  libnl-genl-3-dev,
- libnotcurses-dev (>= 1.4.3),
+ libnotcurses-dev (>= 1.7.4),
  libpcap-dev,
  libreadline-dev,
  libsysfs-dev,
