+++ src/ui/notcurses/channels.h
@@ -1,11 +1,11 @@
-#ifndef OMPHALOS_UI_NCURSES_CHANNELS
-#define OMPHALOS_UI_NCURSES_CHANNELS
+#ifndef OMPHALOS_UI_NOTCURSES_CHANNELS
+#define OMPHALOS_UI_NOTCURSES_CHANNELS
 
 #ifdef __cplusplus
 extern "C" {
 #endif
 
-#include <ui/ncurses/util.h>
+#include <ui/notcurses/util.h>
 
 struct iface_state;
 struct panel_state;
